refactor(common): move UsageErrorPayload next to its exit code
`output.ts` imported the payload type from `cli-errors.ts`, which imports
`outputUsageError` back from `output.ts`. The cycle is type-only today, so it
erases at build time and nothing observes it — but it is one value import in
`output.ts` away from becoming a real module cycle, and the next person to add
one gets to discover that.

`errors.ts` already owns `USAGE_ERROR_CODE`, the constant the payload's
`exit_code` carries, and imports neither module. Moving the interface and its
code union there leaves `cli-errors.ts` (classify) and `output.ts` (write) both
depending on `errors.ts` and no longer on each other's types.

`UsageErrorCode` stays private to its module, re-derived in `cli-errors.ts` as
`UsageErrorPayload["error"]` so the classifier's return type cannot drift from
the field it populates.
